refactor(#73): Move Admin and User List pages to the pages directory and update imports
